RPG beginnings
Added functionality for RPG. Includes a database with tables for players and their stats, and the shop with items.

Also added GM (game master) commands, only accessible to me. These are to be used to initialise the database and for RPG moderation in the future. Can also recreate this for general admin commands in the server.